Updated named network in dynamic scripts

+++ Parser/dynamic_scripts/1_clear_containers.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
-docker network rm streaming_network
+
+net=false
+#delete network
+if $net ; then
+	docker network rm streaming_network
+fi
 #docker container kill streaming_server
 docker container rm streaming_dataset
 docker container rm streaming_server
